+++ src/DevilDaggersInfo.Tools/Ui/Popups/ErrorMessage.cs
@@ -5,16 +5,23 @@ namespace DevilDaggersInfo.Tools.Ui.Popups;
 public class ErrorMessage : Popup
 {
 	private readonly string _errorText;
+	private readonly string? _technicalDetails;
 
-	public ErrorMessage(string id, string errorText)
+	public ErrorMessage(string id, string errorText, string? technicalDetails = null)
 		: base(id)
 	{
 		_errorText = errorText;
+		_technicalDetails = technicalDetails;
 	}
 
 	public override bool Render()
 	{
 		ImGui.TextWrapped(_errorText);
+
+		if (_technicalDetails != null && ImGui.CollapsingHeader("Technical details"))
+		{
+			ImGui.TextWrapped(_technicalDetails);
+		}
 
 		ImGui.Spacing();
 		ImGui.Separator();